+++ src/main/java/com/ch/s1/product/ProductTypeVO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class ProductTypeVO {
 	
-	private int product_type;
+	private long product_type;
 	private String product_type_name; 
 }
