+++ src/main/java/com/ch/s1/seat/SeatVO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class SeatVO {
 	
-	private int seat_num;
+	private long seat_num;
 	private String seat_name;
 }
